[UPDATE] Optimize (#25)
