+++ package.json
@@ -13,12 +13,14 @@
 		"@emotion/styled": "^11.11.0",
 		"@mui/icons-material": "^5.14.7",
 		"@mui/material": "^5.14.7",
+		"@next/bundle-analyzer": "^14.1.0",
 		"@react-google-maps/api": "^2.19.2",
 		"@tanstack/react-query": "^5.13.4",
 		"@types/node": "^20.11.1",
 		"@types/react": "18.2.21",
 		"@types/react-beautiful-dnd": "^13.1.5",
 		"@types/react-dom": "18.2.7",
+		"@types/react-loader-spinner": "^4.0.0",
 		"@types/uuid": "^9.0.5",
 		"@vercel/speed-insights": "^1.0.5",
 		"autoprefixer": "^10.4.16",
@@ -32,12 +34,11 @@
 		"react": "18.2.0",
 		"react-beautiful-dnd": "^13.1.1",
 		"react-dom": "18.2.0",
+		"rtcmulticonnection": "^3.7.1",
 		"socket.io-client": "^4.7.2",
 		"tailwindcss": "3.3.3",
 		"typescript": "5.2.2",
 		"utf-8-validate": "^5.0.10",
-		"uuid": "^9.0.1",
-		"rtcmulticonnection": "^3.7.1",
-		"@types/react-loader-spinner": "^4.0.0"
+		"uuid": "^9.0.1"
 	}
 }
